requirements.txt
web3
python-telegram-bot==20.7
aiohttp
